feat: add XDP program, VLAN support, .rodata port filter
- Add SEC("xdp") entry for pre-stack SYN capture with automatic
  fallback to socket filter when XDP isn't available
- Replace TARGET_PORT array map with volatile .rodata global
  (eliminates per-packet map lookup)
- Add VLAN 802.1Q/QinQ tag peeling in all three program types
- Fix socket filter ignoring VLAN-tagged frames (hardcoded offsets)
- Add EBPF_STATS counter map for ringbuf drop/filter visibility
- Extract shared parse logic into syn_capture.h, deduplicate
  XDP and TC into a single parse_and_submit() inline
- Add -Werror to Dockerfile eBPF build
